fix: Get annotation running
Although the annotation might be running, zero shot learning seems to be
to hard for the model and one shot learning takes too long. I will try
and reduce the size of the examples as well as add an option to classify
only parts of posts as to work faster and serve as a basis to fine tune
tab.